+++ tests/testCorporatePurchase.js
@@ -22,7 +22,7 @@ describe('TestCorporatePurchaseGet', function () {
         let purchases = await starkbank.corporatePurchase.query({ limit: 1 });
         for await (let purchase of purchases) {
             assert(typeof purchase.id == typeof 'string');
-            brcode = await starkbank.corporatePurchase.get(purchase.id);
+            purchase = await starkbank.corporatePurchase.get(purchase.id);
             assert(typeof purchase.id == typeof 'string');
         }
     });